prediction.py
# -*- coding: utf-8 -*-
"""
CSRNET Prediction Module
Adapted for web deployment

Original created on Sat Nov  3 14:46:02 2018
@author: lenovo
"""

import cv2
import h5py
from PIL import Image
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.cm as cm
import os
import logging

# Try to import Keras/TensorFlow
try:
    from keras.models import model_from_json
    KERAS_AVAILABLE = True
except ImportError:
    try:
        from tensorflow.keras.models import model_from_json
        KERAS_AVAILABLE = True
    except ImportError:
        KERAS_AVAILABLE = False
        logging.warning(
            "Neither Keras nor TensorFlow is available. Using mock predictions.")

# Global model variable for caching
_model = None


def load_model():
    """Function to load and return neural network model."""
    global _model

    if _model is not None:
        return _model

    if not KERAS_AVAILABLE:
        logging.warning("Keras/TensorFlow not available, returning None")
        return None

    try:
        model_path = 'models/Model.json'
        weights_path = 'weights/model_B_weights.h5'

        if not os.path.exists(model_path):
            logging.error(f"Model file not found: {model_path}")
            logging.info(
                "Please place your trained Model.json file in the models/ directory")
            return None

        if not os.path.exists(weights_path):
            logging.error(f"Weights file not found: {weights_path}")
            logging.info(
                "Please place your trained model_A_weights.h5 file in the weights/ directory")
            return None

        json_file = open(model_path, 'r')
        loaded_model_json = json_file.read()
        json_file.close()
        loaded_model = model_from_json(loaded_model_json)
        loaded_model.load_weights(weights_path)

        _model = loaded_model
        logging.info("Model loaded successfully")
        return loaded_model

    except Exception as e:
        logging.error(f"Error loading model: {str(e)}")
        return None


def create_img(path):
    """Function to load, normalize and return image."""
    try:
        logging.info(f"Processing image: {path}")
        im = Image.open(path).convert('RGB')

        # Convert to numpy array
        im = np.array(im)

        # Normalize to [0, 1]
        im = im / 255.0

        # Apply ImageNet normalization
        im[:, :, 0] = (im[:, :, 0] - 0.485) / 0.229
        im[:, :, 1] = (im[:, :, 1] - 0.456) / 0.224
        im[:, :, 2] = (im[:, :, 2] - 0.406) / 0.225

        # Add batch dimension
        im = np.expand_dims(im, axis=0)

        logging.info(f"Image processed successfully. Shape: {im.shape}")
        return im

    except Exception as e:
        logging.error(f"Error processing image {path}: {str(e)}")
        raise e


def create_mock_prediction(image):
    """Create a mock prediction when model is not available."""
    # Get image dimensions
    height, width = image.shape[1], image.shape[2]

    # Create a realistic-looking density map
    # Use Gaussian blobs to simulate crowd regions
    # CSRNet typically downsamples by 8x
    density_map = np.zeros((height // 8, width // 8))

    # Add some random Gaussian blobs
    num_blobs = np.random.randint(3, 10)
    total_count = 0

    for _ in range(num_blobs):
        # Random center
        cy = np.random.randint(0, density_map.shape[0])
        cx = np.random.randint(0, density_map.shape[1])

        # Random size and intensity
        sigma_y = np.random.randint(5, 20)
        sigma_x = np.random.randint(5, 20)
        intensity = np.random.uniform(0.5, 2.0)

        # Create Gaussian blob
        y, x = np.ogrid[:density_map.shape[0], :density_map.shape[1]]
        blob = intensity * \
            np.exp(-((x - cx)**2 / (2*sigma_x**2) + (y - cy)**2 / (2*sigma_y**2)))
        density_map += blob

        total_count += intensity * 10  # Rough estimate

    # Add batch dimension to match expected output
    density_map = np.expand_dims(density_map, axis=0)

    return total_count, density_map


def predict(path):
    """Function to load image, predict heat map, generate count and return (count, image, heat map)."""
    try:
        # Load and process image
        image = create_img(path)

        # Load model
        model = load_model()

        if model is None:
            logging.warning("Model not available, using mock prediction")
            # Create mock prediction for demonstration
            count, ans = create_mock_prediction(image)
            logging.info("Mock prediction created")
        else:
            # Make actual prediction
            ans = model.predict(image)
            count = np.sum(ans)
            logging.info(f"Model prediction completed. Count: {count}")

        return count, image, ans

    except Exception as e:
        logging.error(f"Error in predict function: {str(e)}")
        raise e


# Test function for standalone execution
if __name__ == "__main__":
    try:
        # Test with a sample image if available
        test_path = 'test_images/IMG_900.jpg'
        if os.path.exists(test_path):
            ans, img, hmap = predict(test_path)
            print("Predict Count:", ans)

            # Display results
            plt.figure(figsize=(15, 5))

            plt.subplot(1, 3, 1)
            plt.imshow(img.reshape(img.shape[1], img.shape[2], img.shape[3]))
            plt.title('Original Image')
            plt.axis('off')

            plt.subplot(1, 3, 2)
            plt.imshow(hmap.reshape(hmap.shape[1], hmap.shape[2]), cmap=cm.jet)
            plt.title('Density Heat Map')
            plt.colorbar()
            plt.axis('off')

            plt.tight_layout()
            plt.show()
        else:
            print(f"Test image not found: {test_path}")
            print("Please place a test image in the test_images/ directory")
    except Exception as e:
        print(f"Error in test execution: {str(e)}")
