requirements.txt
opencv-python==4.6.0.66
matplotlib==3.5.3
scikit-learn==1.1.3


Flask==3.0.3
flask-cors==4.0.0
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
MarkupSafe==2.1.5


colorama==0.4.6
filelock==3.12.2
fsspec==2023.6.0
greenlet==2.0.2
setuptools==65.5.1

mpmath==1.3.0
sympy==1.12
networkx==2.8.8
pillow==9.4.0

SQLAlchemy==2.0.25

tensorflow==2.10.1        
torch==1.12.1             
torchvision==0.13.1  